Move codes to fetchFromRmp.ts
